Updated Gemfile to use pessimistic locking
Necessary to ensure all dependencies are locked to the minor version so we can more easily update and stay on top of open source changes by using `bundle outdated --only-explicit`, for example, to determine what is outdated and needs to be upgraded.

This was automated by running `pessimize` as provided by the link:https://github.com/joonty/pessimize[Pessimize] gem.