Fixed GitHub Actions to include repository head ref
Necessary for Git Lint to lint commit messages properly since GitHub Actions doesn't provide this by default. Git Lint needs to this determine which commits are on the feature branch.

This includes setting the default branch to `master` since Git Lint assumes `main` by default.

This will help provide consistent commits and, eventually, the generation of release notes which is important to communication changes to the community.