Updated app to require lib folder
Necessary to automatically require all Ruby files, regardless of depth, from the `lib` folder. We could use link:https://github.com/fxn/zeitwerk[Zeitwerk] (and we'll need to soon) but, since we have the Refinements gem, we can use it for now.

There is nothing in the `lib` folder as of yet but will be making use of it soon.