Added image screensaver
Necessary to save a web page as a screenshot. Browser settings are injected with safe defaults and `window_size` is used to resize the browser before updating the DOM and capturing the screenshot because the browser needs to be resized even though the `window_size` should be enough but isn't.

The output of this object will soon be used to create an image suitable for a TRMNL device.