Added Rerun gem
Necessary to automatically reboot the server when making changes locally. This is most useful when working in development and wanting to see changes picked up by the server when working in the UI and/or debugging changes locally. link:https://github.com/alexch/rerun[Documentation].

This gem hasn't been updated in some time so is mostly a temporary solution until finding something better.